test(a11y): move the compact-outline test to its a11y-capable binary
compact_outline_is_smaller_and_every_id_still_resolves lived in the shared
common/mcp_cost.rs module, so it compiled into all eight test binaries. It
launches a fixture with a11y:true, which needs the AT-SPI bus that only the
a11y CI jobs install — so it failed in the X11/Wayland integration jobs
(which run the non-a11y binaries) while passing on a dev box that happens to
have at-spi2-core. Moving it into verification_cost.rs, beside its a11y
siblings and the harness its ignore-message names, scopes it to the one
binary set up to run it. Also drops the now-dead unsafe_code allow from the
shared module.
