+++ crates/glass-testapp/tests/common/mcp_cost.rs
@@ -11,17 +11,9 @@
 //! calls it) — which means the `#[cfg(test)]` unit tests below run under `cargo test
 //! --workspace` regardless.
 
-// One `unsafe { env::set_var }` for pre-spawn GLASS_DISPLAY setup in
-// `compact_outline_is_smaller_and_every_id_still_resolves` (see its `// SAFETY:` note), same
-// opt-out as verification_cost.rs / ignore_regions_e2e.rs / network.rs. It lives here, on this
-// module, rather than on each binary's crate root: this file is compiled into several test
-// binaries (see above) and only this one test needs it.
-#![allow(unsafe_code)]
-
 use std::path::PathBuf;
-use std::time::{Duration, Instant};
+use std::time::Duration;
 
-use glass_core::{AppSpec, AxNodeId, AxTree, Glass, SandboxLevel};
 use glass_mcp::serve::config::ServeConfig;
 use rmcp::model::CallToolRequestParams;
 use rmcp::service::RunningService;
@@ -29,8 +21,6 @@ use rmcp::transport::streamable_http_client::StreamableHttpClientTransportConfig
 use rmcp::transport::StreamableHttpClientTransport;
 use rmcp::{Peer, RoleClient, ServiceExt};
 use serde_json::{json, Value};
-
-use super::Xvfb;
 
 /// Repo root: `crates/glass-testapp` is two levels below it.
 fn repo_root() -> PathBuf {
@@ -143,8 +133,8 @@ pub async fn start_fixture(client: &Peer<RoleClient>) {
 
 /// How long `a11y_outline` retries a snapshot that reports the launched app isn't publishing
 /// an accessibility tree yet.
-const A11Y_SETTLE_TIMEOUT: Duration = Duration::from_secs(5);
-const A11Y_POLL_INTERVAL: Duration = Duration::from_millis(200);
+pub(crate) const A11Y_SETTLE_TIMEOUT: Duration = Duration::from_secs(5);
+pub(crate) const A11Y_POLL_INTERVAL: Duration = Duration::from_millis(200);
 
 /// Snapshot the a11y tree; return the full text (envelope + untrusted outline). Retries for
 /// up to `A11Y_SETTLE_TIMEOUT`: immediately after `glass_start` returns, the launched app's
@@ -375,7 +365,7 @@ impl std::fmt::Display for ArmReport {
 /// tree before giving up. The larger of the two deadlines this consolidates (mcp_cost.rs
 /// previously used 5s, `verification_cost.rs`'s now-removed `poll_until_widgets_present`
 /// used 10s) so the shared function has the more generous headroom of the two under load.
-const WIDGETS_SETTLE_TIMEOUT: Duration = Duration::from_secs(10);
+pub(crate) const WIDGETS_SETTLE_TIMEOUT: Duration = Duration::from_secs(10);
 
 /// Poll (unmetered) until the fixture's three widgets are all present in the a11y tree, or
 /// `WIDGETS_SETTLE_TIMEOUT` elapses (in which case this panics, embedding the last-seen
@@ -633,120 +623,6 @@ pub async fn run_verification_cost(client: &Peer<RoleClient>) -> (ArmReport, Arm
 async fn restart_fixture(client: &Peer<RoleClient>) {
     call(client, "glass_stop", json!({})).await;
     start_fixture(client).await;
-}
-
-/// Sync counterpart to `start_fixture`: launches the same fixture app directly against a
-/// `glass_core::Glass` session instead of over the MCP wire. The wire has no route to the
-/// property this file's cost-and-integrity test needs — `glass_a11y_snapshot` always answers
-/// with `render_compact`'s output (see `glass_mcp::tools::a11y_snapshot`), never the
-/// uncompacted `AxTree::to_outline` render — so that test drives `Glass` directly to see both.
-/// The caller must point `GLASS_DISPLAY` at a live Xvfb before calling this, the same as the
-/// wire-based tests in this suite: `start_on("x11", ...)` attaches to whatever display
-/// `GLASS_DISPLAY` names rather than spawning its own.
-fn start_fixture_sync(glass: &mut Glass) {
-    let (build, run, cwd) = fixture_run_spec();
-    let spec = AppSpec {
-        build: Some(build),
-        run: vec![run],
-        cwd: Some(PathBuf::from(cwd)),
-        env: vec![],
-        window_hint: None,
-        timeout_ms: 120_000, // first egui build is slow, same as start_fixture
-        sandbox: SandboxLevel::Off,
-        a11y: true,
-    };
-    glass
-        .start_on("x11", &spec)
-        .unwrap_or_else(|e| panic!("start_on(x11) failed: {e}"));
-}
-
-/// Sync counterpart to `wait_for_widgets`: polls `Glass::a11y_snapshot` directly rather than
-/// an MCP `Peer`, for the same reason as `start_fixture_sync`. Same retry reasoning as
-/// `wait_for_widgets`/`a11y_outline`: the launched app's toolkit can transiently error (its
-/// AT-SPI subtree not registered yet) or answer with a placeholder root before its widgets are
-/// filled in, so both are retried up to the combined budget.
-fn wait_for_widgets_sync(glass: &mut Glass) -> AxTree {
-    let deadline = Instant::now() + A11Y_SETTLE_TIMEOUT + WIDGETS_SETTLE_TIMEOUT;
-    loop {
-        match glass.a11y_snapshot() {
-            Ok(tree) => {
-                let outline = tree.to_outline();
-                let ready = find_named_button(&outline, "Apply").is_some()
-                    && find_by_role(&outline, "slider").is_some()
-                    && find_by_role(&outline, "text").is_some();
-                if ready {
-                    return tree;
-                }
-                if Instant::now() >= deadline {
-                    panic!(
-                        "wait_for_widgets_sync: timed out waiting for the fixture's widgets; \
-                         last-seen outline:\n{outline}"
-                    );
-                }
-            }
-            Err(e) if Instant::now() < deadline => {
-                let _ = e; // transient during app startup; keep polling
-            }
-            Err(e) => panic!("a11y_snapshot errored: {e}"),
-        }
-        std::thread::sleep(A11Y_POLL_INTERVAL);
-    }
-}
-
-/// Every `#<n>` id at the start of an outline's lines (see `find_by_role`'s doc for the line
-/// shape) — `split_whitespace` + `strip_prefix('#')`, no regex dependency.
-fn ids_in(outline: &str) -> Vec<AxNodeId> {
-    let mut ids = Vec::new();
-    for line in outline.lines() {
-        let Some(tok) = line.split_whitespace().next() else {
-            continue;
-        };
-        let Some(digits) = tok.strip_prefix('#') else {
-            continue;
-        };
-        let Ok(id) = digits.parse::<u32>() else {
-            continue;
-        };
-        ids.push(AxNodeId(id));
-    }
-    ids
-}
-
-/// `render_compact` must remove lines from a real tree (the direction of the change; the exact
-/// ratio is fixture-dependent and deliberately not asserted), and every id it keeps must still
-/// resolve in the full tree — the property that keeps an elided element addressable by
-/// `glass_click_element` / `glass_set_value` after compaction: ids are assigned over the full
-/// tree and compaction must never invent or renumber one.
-#[test]
-#[ignore = "requires an X server + AT-SPI bus; run via scripts/verification-cost.sh"]
-fn compact_outline_is_smaller_and_every_id_still_resolves() {
-    let xvfb = Xvfb::start();
-    // SAFETY: single-threaded test setup; runs before any server task spawns.
-    unsafe { std::env::set_var("GLASS_DISPLAY", &xvfb.display) };
-
-    let mut glass = glass_mcp::boot(None);
-    start_fixture_sync(&mut glass);
-    let tree = wait_for_widgets_sync(&mut glass);
-
-    let full = tree.to_outline();
-    let compact = glass_core::outline::render_compact(&tree);
-
-    assert!(
-        compact.lines().count() < full.lines().count(),
-        "compaction must remove lines (full {} / compact {})",
-        full.lines().count(),
-        compact.lines().count()
-    );
-    for id in ids_in(&compact) {
-        assert!(
-            tree.find(id).is_some(),
-            "#{} appears in the compact outline but resolves to nothing — compaction must \
-             never invent or renumber ids",
-            id.0
-        );
-    }
-
-    let _ = glass.stop();
 }
 
 #[cfg(test)]
